SCSIMODE: configuration file support for mode page settings.
Before a a drive is incorporated into an array we want to be sure that settings
like write-cache-disable and error-recovery-response are set consistently for
each drive in the array.  Add a "SCSIMODE" configuration parameter that:

1/ identifies the disks to apply the policy to by domain name.  When
   domain= is specified multiple times the scsi mode page settings are
   applied to the union of those domains.

2/ Modifies the mode page bytes identified by a tuple of (page, subpage,
   byte).  For now only the '=' operator is supported to assign the
   entire byte.

      _____ page
     /   __ subpage
     | /  _ byte
     | | /
     v v v
     x.y.z=val

Example disable the write cache on all drives in domain0 and domain1:
SCSIMODE domain=domain0 domain=domain1
     # disable write cache
     8.0.2=0
     # enable automatic read/write re-assignment
     1.0.2=0xc0

For multi-byte values specifying something like:
     8.0.0=0
     8.0.1=0
     8.0.2=0
     8.0.3=0
...will result in one write to mode-page-8.  Settings to the same page
are aggregated and the values are set in specified order within a page.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Song Liu <songliubraving@fb.com>